Separate file structure into East / West
