+++ addons/sevenrings/Eastern/CfgVehicles_Eastern.hpp
