+++ addons/sevenrings/Eastern/CfgVehicles_Eastern_Backpacks.hpp
