+++ addons/sevenrings/Eastern/CfgVehicles_Eastern_Vehicles.hpp
