+++ addons/sevenrings/Western/CfgGroups_I.hpp
