+++ addons/sevenrings/Western/CfgVehicles_Western.hpp
