+++ addons/sevenrings/Western/CfgWeapons_Western.hpp
